requirements.txt

pandas
streamlit
matplotlib
python-docx
PyPDF2
requests
beautifulsoup4
